Fixed issue with counter

+++ js/script.js
@@ -1,13 +1,13 @@
 // GET ADD EMPLOYEE FORM AND EMPLOYEE TABLE FROM THE DOM
 let form = document.getElementById('addForm')
 let table = document.getElementById('employees')
+let empCount = document.getElementById('empCount')
 
 // HELPER FUNCTION TO RETURN DOM ELEMENT
 const $ = (id) => document.getElementById(id)
 
 // SET A COUNT VARIABLE TO DISPLAY NEXT TO EMPLOYEES HEADER
-let empCount = $('empCount')
-
+let countValue = 0
 
 // ADD EMPLOYEE
 form.addEventListener('submit', (e) => {
@@ -26,18 +26,25 @@ form.addEventListener('submit', (e) => {
     let row = table.insertRow();
 
     // INSERT A CELL FOR EACH ITEM WITHIN THE NEW ROW
-    let cell1 = row.insertCell(0);
-    let cell2 = row.insertCell(1);
-    let cell3 = row.insertCell(2);
-    let cell4 = row.insertCell(3);
-    let cell5 = row.insertCell(4);
+    let cell1 = row.insertCell(0)
+    let cell2 = row.insertCell(1)
+    let cell3 = row.insertCell(2)
+    let cell4 = row.insertCell(3)
+    let cell5 = row.insertCell(4)
+    let cell6 = row.insertCell(5)
 
     // APPEND THE TEXT VALUES AS TEXT NODES WITHIN THE CELLS
-    cell1.innerHTML = empId
-    cell2.innerHTML = name
-    cell3.innerHTML = ext
-    cell4.innerHTML = email
-    cell5.innerHTML = department
+    let text1 = document.createTextNode(empId)
+    let text2 = document.createTextNode(name)
+    let text3 = document.createTextNode(ext)
+    let text4 = document.createTextNode(email)
+    let text5 = document.createTextNode(department)
+
+    cell1.appendChild(text1)
+    cell2.appendChild(text2)
+    cell3.appendChild(text3)
+    cell4.appendChild(text4)
+    cell5.appendChild(text5)
 
 
     // CREATE THE DELETE BUTTON
@@ -45,7 +52,7 @@ form.addEventListener('submit', (e) => {
     deleteBtn.className = 'btn btn-danger btn-sm float-end delete'
     let textDelete = document.createTextNode('X')
     deleteBtn.appendChild(textDelete)
-    row.appendChild(deleteBtn)
+    cell6.appendChild(deleteBtn)
 
     // RESET THE FORM
     form.reset()
@@ -55,7 +62,8 @@ form.addEventListener('submit', (e) => {
 
 
     // INCREMENENT THE NUMBER OF EMPLOYEES IN THE TABLE
-
+    countValue++
+    empCount.value = '(' + countValue + ')'
     
 
 })
@@ -63,13 +71,15 @@ form.addEventListener('submit', (e) => {
 // DELETE EMPLOYEE
 table.addEventListener('click', (e) => {
     // CHECK AND SEE IF THE DELETE BUTTON WAS CLICKED
-    
+    if (e.target.classList.contains('delete')) {
         // DISPLAY CONFIRMATION OF THE DELETE TO THE USER
         if (confirm('Are you sure you want to delete this record?')) {
             // REMOVE THE SELECTED ROW
-            row = e.target.parentElement.rowIndex
+            row = e.target.parentElement.parentElement.rowIndex
             table.deleteRow(row);
+            countValue--
+            empCount.value = '(' + countValue + ')'
             
         }
-    
+    } 
 })
